+++ src/index.html
@@ -21,7 +21,7 @@
       </div>
       <hr class="border" />
       <div class="tasks-list"></div>
-      <a class="clear" onclick="clearAllCompleted()">Clear all completed</a>
+      <a class="clear">Clear all completed</a>
     </div>
   </body>
 </html>
