+++ src/modules/Constants.js
@@ -1,0 +1,3 @@
+export const tasks = document.querySelector('.tasks-list');
+export const addBtn = document.querySelector('.add');
+export const input = document.querySelector('.add-task');
